+++ ChangeLog.md
@@ -3,6 +3,7 @@
   * Add capability to upload logs to EDDI server
   * Add ability to access EDDI's confguration UI from VoiceAttack
   * Update VoiceAttack profile with fuel tank variables
+  * Provide better translation for VESPER-M4 when speaking that system's name
 
 #2.0.6
   * Update ship information when undocking